Use an explicit list of webkit-patch commands instead of using auto-discovery.
This is a followup to http://crrev.com/1950773002 and related CLs,
which removed remove imports from other __init__.py files; after this,
all of the __init__.py files in webkitpy are empty.

Before this CL, webkit-patch relies on all commands being imported (by
importing the package webkitpy.tool.commands) and then discovering all
subclasses of command using the magic method __subclasses__.

This CL makes it so that commands that are part of webkit-patch are
explicitly listed, and there is no class discovery.

BUG=598897

Review-Url: https://codereview.chromium.org/2070333003
Cr-Original-Commit-Position: refs/heads/master@{#400369}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b5277f6feda049f0941928ae198ed8fed41e15a2