Don't invalidate Inlineboxes being deleted
Invalidation of InlineBoxes being deleted was needed when we used a hash map in
PaintController to track paint invalidations, to ensure a new object allocated
at the same address of an deleted but not invalidated object won't be mistakenly
treated as cached. Now this is unnecessary with client-side cache generations.

BUG=619630
R=chrishtr@chromium.org

Review URL: https://codereview.chromium.org/2063573002 .

Cr-Original-Commit-Position: refs/heads/master@{#400580}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4caf53fcb47c4df2bc26a3b0df8f366a3f4352f7