Let FrameView track object paint invalidations
With client-side invalidation flags, PaintController no longer need to
track invalidation of clients for painting purpose. We also don't need
to distinguish on which GraphicsLayer a client is invalidated because
when it's invalidated it will be invalid on all GraphicsLayers.

Let FrameView track object paint invalidations so that we don't need
to send object paint invalidations to backings.

BUG=617799

Review-Url: https://codereview.chromium.org/2051333005
Cr-Original-Commit-Position: refs/heads/master@{#400605}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5cdeb1b846e2f1c8bcab391e0b0ea55faf5c5b12