Add a release assert to verify the cause of a crash
Other code in LayoutPart checks for the node in addition to the widget
before accessing the widget. It is a proxy for checking whether the
LayoutPart has been destroyed (because the node is cleared in destroy).
Add the check here to see if it hits in the crashing cases.

R=chrishtr@chromium.org
BUG=93805

Review-Url: https://codereview.chromium.org/2083693002
Cr-Original-Commit-Position: refs/heads/master@{#400771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d98e4296e6eab15ff2e56ae93b07de196fc0e37a