Handle pagination breaking on rows correctly when a table has >1 section
When a table has more than one section, layoutRows() needs to have an
up-to-date logicalTop() for the section. Otherwise it will calculate
the wrong offset from the top of the page for the row when figuring out whether
to put a strut on it.

BUG=621359

Review-Url: https://codereview.chromium.org/2076413002
Cr-Original-Commit-Position: refs/heads/master@{#400941}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5aca76371f4c851adb4b1ac398ac45d62073967b