Retain strong reference of BlobCallback in CanvasAsyncBlobCreator
CanvasAsyncBlobCreator holds a BlobCallback and runs it asynchronously.
However, it can be destroyed before the callback invocation, and the
callback doesn't retain its ownership while it's in the task queue.

This CL makes the bound task to retain the shared ownership of
BlobCallback.

BUG=621547

Review-Url: https://codereview.chromium.org/2081873003
Cr-Original-Commit-Position: refs/heads/master@{#401052}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7d195955d69a096ba28bcb4be85df1d64ffa2a8