Set LayoutListBox's height as 'size' attribute * (height of the tallest option).
This will change the current behaviour where we set the size of HTML select element based on the height of first option.
The new behaviour is consistent with other browsers, e.g. Firefox sizes the box as (size * (height of the tallest option)).

BUG=616187

Review-Url: https://codereview.chromium.org/2060493002
Cr-Original-Commit-Position: refs/heads/master@{#401166}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e676eabb8a1e897bbc31a228fe5e49dab06d53e3