Adjust test expectations for infinite-size images.
The test for infinite-size images is expecting the width to be equal to
the maximum possible value of a blink layout value. However, with
use-zoom-for-dsf on and device scale factor !=1 all image widths are
adjusted by the scale factor. So the test's expectation needs to be
changed because the effective maximum value is now different.

BUG=622000

Review-Url: https://codereview.chromium.org/2085153002
Cr-Original-Commit-Position: refs/heads/master@{#401397}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e691b41f4226ecb00323a46fedb97d0ae2b67047