Use unfragmented clip bounds when fragmentation is inhibited due to compositing
This CL fixes one case we missed to patch up in:
https://codereview.chromium.org/2028893002

When a PaintLayer has opacity it will use the (incorrectly) fragmented bounds
for CompositingRecorder to save Skia layer. The bound is soft so it is not
always reproducible depends on internal implementation.

Review-Url: https://codereview.chromium.org/2088333002
Cr-Original-Commit-Position: refs/heads/master@{#401504}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 533c7f586be0bb8fc64705a82cdc8133d455b17f