Fix positioning of CSS reference clips in composited layers.
When painting into composited scrolling layers, clip paths must use
an offset from the root location to account for the origin of the
the layer's paint coordinate system. SVG clips referenced by CSS,
with the default UserSpaceOnUse coordinate system were not applying
this offset as required. When the clip is in ObjectBoundingBox
coordinates the offset was applied as a side effect of the object's
bounding box.

This patch corrects the position by adding the necessary offset when
the clip coordinate system is UserSpaceOnUse.

R=fs@opera.com
BUG=617265

Review-Url: https://codereview.chromium.org/2078343005
Cr-Original-Commit-Position: refs/heads/master@{#401642}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6cc64ba85ee2f7927578efb8d91641db367c2f0