cc: Make custom scrollbar scroll on main thread
The compositor hit testing needs to let the main thread deal with
scrolling on a scrollbar layer. However some webpages on high-DPI screen
have a custom scrollbar. The compositor cannot figure out if a layer is
a custom scrollbar as it is only a GraphicsLayer in blink.

This CL forces custom scrollbars scroll on main thread so that gestures
on scrollbars can be correctly processed.

BUG=618183
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/2085853003
Cr-Original-Commit-Position: refs/heads/master@{#401891}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 599a0d9d885bae01856dc0c716bf851389fe3330