Reland: WTF: Simplify RefPtr::operator bool.
The problematic case in String::find on ARM should now be gone, and so this
should be both simpler and equally fast.

Comparing to nullptr is used to coerce to boolean to work around bug 621177
(even though this is usually not the preferred way to do this in Blink).

BUG=607208,621177

Review-Url: https://codereview.chromium.org/2090973002
Cr-Original-Commit-Position: refs/heads/master@{#401897}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4db41c4c92314447a07a2255d97c344547f97ce7