Fix http-to-https-redirect-and-register.html for --site-per-process.
The test was assuming that the onload callback would be dispatched
before the postMessage from SW registration, which wasn't happening
when the frame in the second promise_test was an OOPIF.  This CL
changes the onload and postMessage handlers to be independent of
ordering.  See bug for more details.

BUG=622887

Review-Url: https://codereview.chromium.org/2092953003
Cr-Original-Commit-Position: refs/heads/master@{#401900}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 033dafcf829151ce612ed36460049e0a282fbf38