Disable drawing of video backgrounds in OverlayFullscreenVideo
This is a follow up to http://crrev.com/1907753002 to fix cases where the
video background is now visible in OverlayFullscreenVideo, where it's not
desired. This forcefully stops the GraphicsLayer associated with the
video from drawing its content (in particular, the background) when OFV
is enabled.

TEST=LayoutTests, manual testing
BUG=615245, 603521

Review-Url: https://codereview.chromium.org/2083223002
Cr-Original-Commit-Position: refs/heads/master@{#401916}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 014e89fdbe83fd895898331249f9bba28cc71130