[OriginTrials] Raise an exception at compile time for OT-enabled methods
We don't currently support origin trials on individual methods (methods which
are part of an origin-trial-enabled attribute are supported), and the
generated bindings code for these is incorrect. This patch causes an
exception to be raised at compile time to ensure that noone tries to add
the OriginTrialEnabled attribute to a method.

This also prevents trials from being defined on:
 - Iterators
 - Maplike and Setlike methods
 - Stringifiers and Serializers

As part of this patch, the methods defined in TestObject.idl which were
generating incorrect code have been commented out.

BUG=621641
R=chasej@chromium.org, haraken@chromium.org

Review-Url: https://codereview.chromium.org/2090773005
Cr-Original-Commit-Position: refs/heads/master@{#401945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0fa26a48c0341e8676aaafc1f33fc825d3581709