Fix crash when handling GestureScrollBegin and can't get a target node.
In r401997 I speculatively removed the early return if
m_scrollGestureHandlingNode was NULL. It turns out this can happen, I'm guessing
during initialization/destruction, the documentElement is NULL so we can reach
this state. This patch fixes the crash by refusing to handle the event if we
can't get a node to scroll.

BUG=623460

Review-Url: https://codereview.chromium.org/2098333002
Cr-Original-Commit-Position: refs/heads/master@{#402177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d855b4a44ca015cc874b69b828717bb4293c3092