Make WTF::Function convertible to base::Callback
WTF::Function is now a wrapper of base::Callback, and has a base::Callback
as a member.
This CL adds an explicit conversion operator to WTF::Function to allow
accessing the internal base::Callback, and use it where WTF::Function was
converted to base::Callback manually.

Review-Url: https://codereview.chromium.org/2099203002
Cr-Original-Commit-Position: refs/heads/master@{#402421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a942060fcc0b8f030b5fa916a931f790d2183b16