Fix the timeout in texImage-imageBitmap-from-image.html
This test is timed out some times, and the reason is that the test assumes
two source images will be loaded in certain order, where in reality it
may not be the case.

This CL fixes it by using promises when loading an image source.

Review-Url: https://codereview.chromium.org/2093133002
Cr-Original-Commit-Position: refs/heads/master@{#402793}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6eca95b98826580235a1cdb2338f076381424a03