Revert of Reland "Replace ICU with CED for auto encoding detection" (patchset #1 id:1 of https://codereview.chromium.org/2105263002/ )
Reason for revert:
Sheriff is reverting this patchset on suspicion of
adding a new static constructor (string) causing a failure of the "sizes" step
of the build.

Link to broken build:
https://build.chromium.org/p/chromium/builders/Linux%20x64/builds/21618

Relevant log lines:
FAILED linux-release-64/sizes/chrome-si/initializers: actual 9, expected 7,
better lower

We think this is ultimately coming from encodings.cc which is part of the third party ced package, and might be related to this DEPS roll.
revert_cq: 1
revert_reason_textarea: A revert of this CL (patchset #1 id:1) has been created in
https://codereview.chromium.org/2110853002/ by petewil@chromium.org.

The reason for reverting is: Sheriff is reverting this patchset on suspicion of
adding a new static constructor (string) causing a failure of the "sizes" step
of the build.

Link to broken build:
https://build.chromium.org/p/chromium/builders/Linux%20x64/builds/21618

Relevant log lines:
FAILED linux-release-64/sizes/chrome-si/initializers: actual 9, expected 7,
better lower

We think this is ultimately coming from encodings.cc which is part of the third party ced package, and might be related to this DEPS roll.

Original issue's description:
> Reland "Replace ICU with CED for auto encoding detection"
>
> The patch was reverted due to broken build on Windows 8 gyp.
> gyp build was missing a compiler option suppressing a warning
> message while gn build already handles it. crrev.com/2103793005
> now fixed the issue.
>
> This reverts commit e6ea446fa4e5628f61f35e5fe88e40cbb0d0b924.
>
> BUG=597488
>
> Committed: https://crrev.com/0c155762b0654045ed3031e5d1e2b447e40ca42e
> Cr-Commit-Position: refs/heads/master@{#402832}

TBR=tkent@chromium.org,jinsukkim@chromium.org
# Skipping CQ checks because original CL landed less than 1 days ago.
NOPRESUBMIT=true
NOTREECHECKS=true
NOTRY=true
BUG=597488

Review-Url: https://codereview.chromium.org/2110833005
Cr-Original-Commit-Position: refs/heads/master@{#402901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dbd9f219ef9b615d9e116964e46b1a390c9354a1